docs(forms): list all sub-components and API in extended features
