+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/create-component/ListBasisAPIs.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useStaticQuery, graphql } from 'gatsby'
 import ListSummaryFromEdges from '../../../../../shared/parts/ListSummaryFromEdges'
 
-export default function ListBasisAPIs() {
+export default function ListBasisAPIs(props) {
   const {
     allMdx: { edges },
   } = useStaticQuery(graphql`
@@ -35,5 +35,5 @@ export default function ListBasisAPIs() {
     }
   `)
 
-  return <ListSummaryFromEdges edges={edges} />
+  return <ListSummaryFromEdges edges={edges} {...props} />
 }
