+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/SubmitIndicator.mdx
@@ -1,7 +1,6 @@
 ---
 title: 'SubmitIndicator'
 description: '`Form.SubmitIndicator` lets you show an indicator while async form operations are performed.'
-componentType: 'advanced-api'
 hideInMenu: true
 showTabs: true
 tabs:
