+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/Visibility.mdx
@@ -1,7 +1,6 @@
 ---
 title: 'Visibility'
 description: '`Form.Visibility` makes it possible to hide components and elements on the screen based on the dynamic state of data.'
-componentType: 'advanced-api'
 hideInMenu: true
 showTabs: true
 tabs:
