+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/components.mdx
@@ -2,8 +2,8 @@
 showTabs: true
 ---
 
-import ListComponents from './ListComponents'
+import ListFormComponents from './ListFormComponents'
 
 ## Components
 
-<ListComponents />
+<ListFormComponents />
