+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/error-messages.mdx
@@ -3,6 +3,7 @@ title: 'Error messages'
 description: 'Error messages in Eufemia Forms are used to provide feedback to users when there are issues with their input.'
 hideInMenu: true
 showTabs: true
+componentType: 'docs'
 tabs:
   - title: Info
     key: '/info'
