+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/getData.mdx
@@ -1,7 +1,6 @@
 ---
 title: 'getData'
 description: '`Form.getData` lets you access your form data outside of the form context.'
-componentType: 'advanced-api'
 hideInMenu: true
 showTabs: true
 tabs:
