+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/schema-validation.mdx
@@ -3,6 +3,7 @@ title: 'Schema validation'
 description: 'Schema validation can be done with a JSON Schema which makes it possible to describe the data structure and validation needs, both for the individual value, and more complex rules across the data set.'
 hideInMenu: true
 showTabs: true
+componentType: 'docs'
 tabs:
   - title: Info
     key: '/info'
