+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/setData.mdx
@@ -1,7 +1,6 @@
 ---
 title: 'setData'
 description: '`Form.setData` lets you set or modify your form data outside of the form context.'
-componentType: 'advanced-api'
 hideInMenu: true
 showTabs: true
 tabs:
