+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/useData.mdx
@@ -1,7 +1,6 @@
 ---
 title: 'useData'
 description: '`Form.useData` lets you access or modify your form data outside of the form context within your application.'
-componentType: 'advanced-api'
 hideInMenu: true
 showTabs: true
 tabs:
