+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Form/useError.mdx
@@ -1,7 +1,6 @@
 ---
 title: 'useError'
 description: '`Form.useError` lets you monitor your form errors outside of the form itself.'
-componentType: 'advanced-api'
 hideInMenu: true
 showTabs: true
 tabs:
