+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Iterate.mdx
@@ -1,7 +1,7 @@
 ---
 title: 'Iterate'
 description: '`Iterate` is components and functionality for traversing values and parts of data sets such as arrays, which contain a varying number of elements where the number of components on the screen depends on how many elements the data consists of.'
-componentType: 'advanced-api'
+componentType: 'iterate-api'
 hideInMenu: true
 showTabs: true
 tabs:
