+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/Iterate/components.mdx
@@ -2,8 +2,8 @@
 showTabs: true
 ---
 
-import ListComponents from './ListComponents'
+import ListIterateComponents from './ListIterateComponents'
 
 ## Components
 
-<ListComponents />
+<ListIterateComponents />
