+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/extended-features/ListAdvancedAPIs.tsx
@@ -1,19 +1,26 @@
 import React from 'react'
 import { useStaticQuery, graphql } from 'gatsby'
-import ListSummaryFromEdges from '../../../../../../shared/parts/ListSummaryFromEdges'
+import ListSummaryFromEdges from '../../../../../shared/parts/ListSummaryFromEdges'
 
-export default function ListComponents(props) {
+/**
+ * As of now, this component is not used in the portal.
+ * All the advanced APIs are listed in the `extended-features.mdx` file, manually.
+ */
+
+export default function ListAdvancedAPIs(props) {
   const {
     allMdx: { edges },
   } = useStaticQuery(graphql`
     {
       allMdx(
         filter: {
-          frontmatter: { title: { ne: null }, draft: { ne: true } }
+          frontmatter: {
+            title: { ne: null }
+            draft: { ne: true }
+            componentType: { in: "advanced-api" }
+          }
           internal: {
-            contentFilePath: {
-              glob: "**/uilib/extensions/forms/extended-features/Form/**/*"
-            }
+            contentFilePath: { glob: "**/uilib/extensions/forms/**/*" }
           }
         }
         sort: [
